Move globus code into one file, ready to become a staging provider
There will be one Globus object per GlobusScheme now, rather than one per data manager - this is
to move globus-related state out of the data manager, in preparation for future PRs where
staging providers manage their own configuration, with the data manager knowing nothing other than
a reference to the staging provider.

I think this shouldn't cause significantly-noticeable change in behaviour.

Tested on my laptop with stage-in and stage-out testing.
