+++ parsl/data_provider/data_manager.py
@@ -1,12 +1,9 @@
-import os
 import logging
 import concurrent.futures as cf
-from parsl.data_provider.scheme import GlobusScheme
 from parsl.executors.base import ParslExecutor
 from parsl.data_provider.ftp import _ftp_stage_in_app
-from parsl.data_provider.globus import get_globus
+from parsl.data_provider.globus import _get_globus_scheme
 from parsl.data_provider.http import _http_stage_in_app
-from parsl.app.app import python_app
 
 logger = logging.getLogger(__name__)
 
@@ -64,37 +61,6 @@ class DataManager(ParslExecutor):
     def scaling_enabled(self):
         return self._scaling_enabled
 
-    def initialize_globus(self):
-        if self.globus is None:
-            self.globus = get_globus()
-
-    def _get_globus_endpoint(self, executor_label):
-        if executor_label is None:
-            raise ValueError("executor_label is mandatory")
-        executor = self.dfk.executors[executor_label]
-        if not hasattr(executor, "storage_access"):
-            raise ValueError("specified executor does not have storage_access attribute")
-        for scheme in executor.storage_access:
-            if isinstance(scheme, GlobusScheme):
-                if executor.working_dir:
-                    working_dir = os.path.normpath(executor.working_dir)
-                else:
-                    raise ValueError("executor working_dir must be specified for GlobusScheme")
-                if scheme.endpoint_path and scheme.local_path:
-                    endpoint_path = os.path.normpath(scheme.endpoint_path)
-                    local_path = os.path.normpath(scheme.local_path)
-                    common_path = os.path.commonpath((local_path, working_dir))
-                    if local_path != common_path:
-                        raise Exception('"local_path" must be equal or an absolute subpath of "working_dir"')
-                    relative_path = os.path.relpath(working_dir, common_path)
-                    endpoint_path = os.path.join(endpoint_path, relative_path)
-                else:
-                    endpoint_path = working_dir
-                return {'endpoint_uuid': scheme.endpoint_uuid,
-                        'endpoint_path': endpoint_path,
-                        'working_dir': working_dir}
-        raise Exception('No suitable Globus endpoint defined for executor {}'.format(executor_label))
-
     def stage_in(self, file, executor):
         """Transport the file from the input source to the executor.
 
@@ -117,28 +83,28 @@ class DataManager(ParslExecutor):
             app_fut = stage_in_app(working_dir, outputs=[file])
             return app_fut._outputs[0]
         elif file.scheme == 'globus':
-            globus_ep = self._get_globus_endpoint(executor)
-            stage_in_app = self._globus_stage_in_app()
-            app_fut = stage_in_app(globus_ep, outputs=[file])
+            # what should happen here is...
+            # we should acquire the GlobusScheme object that goes with
+            # this executor (rather than _get_globus_endpoint)
+            # and then ask the scheme to provide the stage_in_app without
+            # invocations to that app needing any further globus specific
+            # parameters.
+            # The longer term path is then that http/ftp also become a
+            # Scheme, and we allow each one a chance to inspect the
+            # file to see if it can handle it (rather than the data manager
+            # having to know about staging)
+            # At that point, we wouldn't be looking up the GlobusScheme
+            # object explicitly - instead we'd be in possession of it just
+            # like the other schemes... and each scheme would know all the
+            # parameterisation it needs to.
+            # so the three real cases of this if/elif block should
+            # look pretty much identical
+            globus_scheme = _get_globus_scheme(self.dfk, executor)
+            stage_in_app = globus_scheme._globus_stage_in_app(executor=executor, dfk=self.dfk)
+            app_fut = stage_in_app(outputs=[file])
             return app_fut._outputs[0]
         else:
             raise Exception('Staging in with unknown file scheme {} is not supported'.format(file.scheme))
-
-    def _globus_stage_in_app(self):
-        return python_app(executors=['data_manager'], data_flow_kernel=self.dfk)(self._globus_stage_in)
-
-    def _globus_stage_in(self, globus_ep, outputs=[]):
-        file = outputs[0]
-        file.local_path = os.path.join(
-                globus_ep['working_dir'], file.filename)
-        dst_path = os.path.join(
-                globus_ep['endpoint_path'], file.filename)
-
-        self.initialize_globus()
-
-        self.globus.transfer_file(
-                file.netloc, globus_ep['endpoint_uuid'],
-                file.path, dst_path)
 
     def stage_out(self, file, executor):
         """Transport the file from the local filesystem to the remote Globus endpoint.
@@ -156,22 +122,8 @@ class DataManager(ParslExecutor):
         elif file.scheme == 'ftp':
             raise Exception('FTP file staging out is not supported')
         elif file.scheme == 'globus':
-            globus_ep = self._get_globus_endpoint(executor)
-            stage_out_app = self._globus_stage_out_app()
-            return stage_out_app(globus_ep, inputs=[file])
+            globus_scheme = _get_globus_scheme(self.dfk, executor)
+            stage_out_app = globus_scheme._globus_stage_out_app(executor=executor, dfk=self.dfk)
+            return stage_out_app(inputs=[file])
         else:
             raise Exception('Staging out with unknown file scheme {} is not supported'.format(file.scheme))
-
-    def _globus_stage_out_app(self):
-        return python_app(executors=['data_manager'], data_flow_kernel=self.dfk)(self._globus_stage_out)
-
-    def _globus_stage_out(self, globus_ep, inputs=[]):
-        file = inputs[0]
-        src_path = os.path.join(globus_ep['endpoint_path'], file.filename)
-
-        self.initialize_globus()
-
-        self.globus.transfer_file(
-            globus_ep['endpoint_uuid'], file.netloc,
-            src_path, file.path
-        )
